% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Assemble microdata across surveys}
\usage{
assemble(year, var, respondent = "household", M = 1, cores = 1)
}
\arguments{
\item{year}{Integer. One or more years for which to return results (i.e. the ACS recipient year).}

\item{var}{Character. Name of one or more variables to return. May contain household- and/or person-level variables. See Details.}

\item{respondent}{Character. Should household or person-level microdata be returned?}

\item{M}{Integer. The first \code{M} implicates are returned for fused variables. Set \code{M = Inf} to return all implicates. Ignored if \code{var} contains only ACS variables (i.e. no implicates)}

\item{cores}{Integer. Number of cores used by the \code{\link[fst]{fst-package}} when reading from disk.}
}
\value{
A keyed data table containing the following columns, in addition to the variables named in \code{var}:
\describe{
\item{M}{Implicate number. See \code{\link{fuse}}.}
\item{year}{Year of the ACS recipient microdata.}
\item{hid}{ACS household ID using fusionACS convention.}
\item{pid}{ACS person ID using fusionACS convention, if \code{respondent = "person"}.}
\item{weight}{ACS microdata primary sample weight.}
}
}
\description{
For fusionACS usage only. Provides a safe and efficient way to assemble (merge) fused microdata across surveys to return a single data table with the requested variables. The requested variables can come from any fused (donor) survey and/or the American Community Survey (ACS). The necessary variables are automatically and efficiently read from the appropriate local file and safely merged on household and/or person ID variables, optionally collapsing or expanding records as necessary depending on the \code{respondent} argument. Assumes (and checks for) a local \code{/fusionData} directory with appropriate file structure and conventions.
}
\details{
The \code{var} argument can contain a mix of household- and/or person-level variables. When \code{respondent = "household"}, the reference person (i.e. head of household) value is returned for any person-level variables. When \code{respondent = "person"}, the values of any household-level variables are replicated for each person in the household.

A warning is issued if any \code{var} cannot be located in the available local files.
}
\examples{
# NOTE: Requires local /fusionData directory containing the necessary ACS and .fsd files
test <- assemble(year = 2018:2019,
                 var = c("dollarel", "hincp", "agep", "state"),
                 respondent = "household",
                 M = 1)
}
