% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two analyses of synthetic data}
\usage{
compare(analysis1, analysis2)
}
\arguments{
\item{analysis1}{Results tibble from \link{analyze()}.}

\item{analysis2}{Results tibble from \link{analyze()}.}
}
\value{
A tibble reporting the \code{overlap} coefficient and associated confidence intervals of each statistic.
}
\description{
Calculates confidence interval overlap coefficients for comparing two results from \link{analyze()}.
}
\details{
Using the 95\% confidence intervals for each statistic common to both \code{analysis1} and \code{analysis2}, the overlap coefficient is calculated as proposed by Karr et al. (2006). Overlap equals one if the two confidence intervals are identical, zero if the intervals barely touch, and is increasingly negative as two confidence intervals move further apart.
}
\examples{
# See usage in Examples of ?analyze().
}
\references{
A. F Karr, C. N Kohnen, A Oganian, J. P Reiter & A. P Sanil. (2006). A Framework for Evaluating the Utility of Data Altered to Protect Confidentiality, The American Statistician, 60:3, 224-232.
}
