% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuseCART.R
\name{fuseCART}
\alias{fuseCART}
\title{Fuse variables to a recipient dataset using CART fusion model}
\usage{
fuseCART(data, train.object, ignore_self = FALSE)
}
\arguments{
\item{data}{Data frame. Recipient dataset. All categorical variables should be factors and ordered whenever possible. Data types and levels are strictly validated against predictor variables defined in \code{train.object}.}

\item{train.object}{Output from a successful call to \link{train}.}

\item{ignore_self}{Logical. If \code{TRUE}, the simulation step excludes "self matches" (i.e. row 1 in \code{data} cannot match with row 1 in the original donor. Only useful for validation exercises. Do not use otherwise.}
}
\value{
For \code{fuse()}, a data frame with same number of rows as \code{data} and one column for each synthetic fusion variable. The order of the columns reflects the order in which they where fused.

For \code{fuseM()}, a data frame with number of rows equal to \code{M * nrow(data)}. Integer column ".M" indicates implicate assignment of each observation. Note that the ordering of recipient observations is consistent within implicates, so do not change the row order if using with \code{analyze()}.
}
\description{
Fuse variables to a recipient dataset using a fusion model object produced by \code{train()}. \code{fuseM()} provides a convenience wrapper for generating multiple implicates.
}
\examples{
# Build a fusion model using RECS microdata
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fit <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Generate single implicate of synthetic 'fusion.vars',
#  using original RECS data as the recipient
recipient <- recs[predictor.vars]
sim <- fuse(data = recipient, train.object = fit)
head(sim)

# Generate multiple implicates
sim <- fuseM(data = recipient, train.object = fit, M = 5)
head(sim)
table(sim$.M)
}
