% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionModel-package.R
\docType{package}
\name{fusionModel-package}
\alias{fusionModel}
\alias{fusionModel-package}
\title{fusionModel: Data fusion and analysis of synthetic data in R}
\description{
Data fusion and analysis of synthetic data in R.
}
\author{
\strong{Maintainer}: Kevin Ummel \email{ummel@berkeley.edu}

}
\keyword{internal}
