% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescreen.R
\name{prescreen}
\alias{prescreen}
\title{Fast initial screening of predictor variables}
\usage{
prescreen(
  data,
  y,
  x,
  weight = NULL,
  fraction = 1,
  cor_thresh = 0.02,
  lasso_thresh = 0.95,
  cores = 1
)
}
\arguments{
\item{data}{Data frame. Training dataset. All categorical variables should be factors and ordered whenever possible.}

\item{y}{Character. Response variable(s).}

\item{x}{Character. Predictor variables.}

\item{weight}{Character. Name of the observation weights column in \code{data}. If NULL (default), uniform weights are assumed.}

\item{fraction}{Numeric. Fraction of observations in \code{data} to randomly sample. For larger datasets, sampling often has minimal effect on results but speeds up computation.}

\item{cor_thresh}{Numeric. Predictors that exhibit less than \code{cor_thresh} absolute correlation with a \code{y} variable are screened out prior to the LASSO step.}

\item{lasso_thresh}{Numeric. Controls how aggressively the LASSO step screens out predictors. Lower value is more aggressive. \code{lasso_thresh = 0.95}, for example, retains predictors that collectively explain at least 95\% of the deviance explained by a "full" model.}

\item{cores}{Integer. Number of cores used. Only applicable on Unix systems.}
}
\value{
Character vector giving the \code{x} variables that passed both the correlation and LASSO screens when predicting at least one of the \code{y} variables.
}
\description{
Quickly detect which predictor variables in a proposed analysis are likely to be consequential in more detailed models. Useful when there are many predictor variables and not all are likely to be material to subsequent model-fitting. \code{prescreen()} employs an absolute correlation screen and then a LASSO model to identify a plausible predictor subset for each response variable.
}
\examples{
?recs
fusion.vars <- names(recs)[13:18]
predictor.vars <- names(recs)[2:12]
xretain <- prescreen(data = recs, y = fusion.vars, x = predictor.vars)
xretain

}
