% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recs}
\alias{recs}
\title{Data from the 2015 Residential Energy Consumption Survey (RECS)}
\format{
A tibble (\code{tbl_df}) with 5,686 rows and 124 variables:
\describe{
\item{weight}{Primary sampling weight.}
\item{income}{Annual gross household income for the last year.}
\item{age}{Respondent age.}
\item{race}{Respondent race.}
\item{education}{Highest education completed by respondent.}
\item{employment}{Respondent employment status.}
\item{hh_size}{Number of household members.}
\item{division}{Census Division.}
\item{urban_rural}{Census 2010 Urban Type.}
\item{climate}{IECC Climate Code.}
\item{renter}{Is household renting the home?}
\item{home_type}{Type of housing unit.}
\item{year_built}{Range when housing unit was built.}
\item{square_feet}{Total square footage.}
\item{insulation}{Level of insulation.}
\item{heating}{Main space heating fuel.}
\item{aircon}{Type of air conditioning equipment used.}
\item{centralac_age}{Age of central air conditioner.}
\item{televisions}{Number of televisions used.}
\item{disconnect}{Frequency of receiving disconnect notice.}
\item{electricity}{Total annual electricity usage, in kilowatthours.}
\item{natural_gas}{Total annual natural gas usage, in hundred cubic feet.}
\item{fuel_oil}{Total annual fuel oil/kerosene usage, in gallons.}
\item{propane}{Total annual propane usage, in gallons.}
\item{propane_btu}{Total annual propane usage, in thousand Btu.}
\item{propane_expend}{Total annual propane expenditure, in dollars.}
\item{heating_share}{Share of total energy used for space heating.}
\item{cooling_share}{Share of total energy used for cooling (AC and fans).}
\item{other_share}{Share of total energy used for other end-uses.}
\item{use_ng}{Logical indicating if household uses natural gas.}
\item{have_ac}{Logical indicating if household has air conditioning.}
\item{rep_1, rep_2, ..., rep_96}{Replicate weights (rep_1 through rep_96) for uncertainty estimation.}
}
}
\usage{
recs
}
\description{
Pre-processed, household-level microdata containing a selection of 31
variables derived from the 2015 RECS, plus 96 survey replicate weights.
A variety of data types are included (numeric, factor, logical). There are no
missing values (\code{NA}). Variable names have been standardized from the original
EIA microdata format.
}
\examples{
data(recs)
head(recs)
}
\references{
U.S. Energy Information Administration (EIA). (2015). \emph{Residential Energy Consumption Survey (RECS)}.
\url{https://www.eia.gov/consumption/residential/data/2015/}
}
\keyword{datasets}
