% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze synthetic data to calculate point estimates and uncertainty}
\usage{
analyze(
  formula,
  by,
  synthetic,
  static,
  donor.N,
  syn.N = nrow(static),
  var.scale = 4,
  cores = 1
)
}
\arguments{
\item{formula}{A formula expression describing the desired analysis. Example: \code{var1 ~ 1} requests mean and totals if \code{var1} is numeric or proportions if \code{var1} is a factor.}

\item{by}{Character. Column name(s) in \code{synthetic} or \code{static} (typically factors) that collectively define the population subgroups to analyze.}

\item{synthetic}{List of data frames containing synthetic implicates, each with the same number of rows as \code{static}. Typically generated by \link{fuseM}.}

\item{static}{Data frame containing static (non-synthetic) variables with the same number of rows as \code{synthetic}. These data are "static" since they do not vary across implicates. At a minimum, this must include a column named \code{"weight"} giving the primary sampling weights and at least two columns of replicate weights denoted \code{"rep_1"}, \code{"rep_2"}, etc.}

\item{donor.N}{Number of (un-weighted) observations in the original "donor" data used to generate the implicates.}

\item{syn.N}{Number of (un-weighted) observations in the "recipient" data. Usually \code{nrow(static)}, the default.}

\item{var.scale}{Value by which to scale the standard variance calculated across replicate weights. This is specified by the survey design. \code{var.scale = 4} (the default) for both RECS and ACS.}

\item{cores}{Integer. Number of cores used for parallel operations.}
}
\value{
A tibble reporting results for the \code{response} variable across population subgroups of \code{by}. The returned quantities are:

\describe{
\item{estimate}{Point estimate. Mean of estimates across the implicates.}
\item{std_error}{Pooled standard error of the point estimate.}
\item{lower_ci}{Lower bound of 95\% confidence interval.}
\item{upper_ci}{Upper bound of 95\% confidence interval.}
\item{degf}{Degrees of freedom of t-distribution if calculating custom confidence intervals.}
}
}
\description{
Calculation of point estimates and associated standard errors for analyses using synthetic microdata. Analysis is currently limited to means, totals, and proportions, (optionally) calculated for population subgroups. Plan to extend to regression model coefficients and custom microsimulation analyses in future. User must supply a list of synthetic implicates (typically generated by \link{fuseM}) and "static" (non-synthetic) variables required by the analysis, including sampling and replicate weights.
}
\details{
The requested analysis is performed separately for each implicate. The final point estimate is the mean estimate across implicates. The final standard error is pooled across implicates using Reiter's pooling rules for partially synthetic data (Reiter, 2003). The estimate and standard error of each implicate is calculated using the provided observation weights. Calculations utilize custom code leveraging \code{\link[data.table]{data.table}} operatons for speed and memory efficiency.

The within-implicate variance is calculated around the mean of the replicates, rather than the point estimate. This is equivalent to \code{mse = FALSE} in \code{\link[survey]{svrepdesign}}. While this is not necessarily correct for all surveys, the effect on final errors is negligible.
}
\examples{
# Build a fusion model using RECS microdata
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fit <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Generate 5 implicates of synthetic 'fusion.vars',
#  using original RECS data as the recipient
sim <- fuseM(data = recs, train.object = fit, M = 5)

# Analyze electricity consumption, by climate zone
result <- analyze(formula = electricity ~ 1,
                  by = "climate",
                  synthetic = sim,
                  static = recs,
                  donor.N = nrow(recs))

# Analyze air conditioning technology, by race
result <- analyze(formula = aircon ~ 1,
                  by = "race",
                  synthetic = sim,
                  static = recs,
                  donor.N = nrow(recs))
}
\references{
Reiter, J.P. (2003). Inference for Partially Synthetic,
Public Use Microdata Sets. \emph{Survey Methodology}, \bold{29}, 181-189.
}
