% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze synthetic data to calculate point estimates and uncertainty}
\usage{
analyze(formula, by, synthetic, static, design, donor.N, cores = 1)
}
\arguments{
\item{formula}{A formula expression describing the desired analysis. Example: \code{var1 ~ 1} requests mean and totals if \code{var1} is numeric or proportions if \code{var1} is a factor.}

\item{by}{Character. Column name(s) in \code{synthetic} or \code{static} (typically factors) that collectively define the population subgroups to analyze.}

\item{synthetic}{List of data frames containing synthetic implicates, each with the same number of rows as \code{static}. Typically generated by \link{fuseM}.}

\item{static}{Data frame containing static (non-synthetic) variables with the same number of rows as \code{synthetic}. These data are "static" since they do not vary across implicates.}

\item{design}{Object of class \code{"svyrep.design"} generated by \code{\link[survey]{svrepdesign}}. See Examples.}

\item{donor.N}{Number of (un-weighted) observations in the original "donor" data used to generate the implicates.}

\item{cores}{Integer. Number of cores used for parallel operations.}
}
\value{
A tibble reporting results for the \code{response} variable across population subgroups of \code{by}. The returned quantities are:

\describe{
\item{estimate}{Point estimate. Mean of estimates across the implicates.}
\item{std_error}{Pooled standard error of the point estimate.}
\item{lower_ci}{Lower bound of 95\% confidence interval.}
\item{upper_ci}{Upper bound of 95\% confidence interval.}
\item{degf}{Degrees of freedom of t-distribution if calculating custom confidence intervals.}
}
}
\description{
Calculation of point estimates and associated standard errors for analyses using synthetic microdata. Analysis is currently limited to means, totals, and proportions, (optionally) calculated for population subgroups. Will extend to regression model coefficients in future. User must supply a list of synthetic implicates (typically generated by \link{fuseM}), as well as any "static" (non-synthetic) variables required by the analysis and a survey "design" object containing observation weights and information about how to employ replicate weights (see \code{\link[survey]{svrepdesign}}).
}
\details{
The requested analysis is performed separately for each implicate. The final point estimate is the mean estimate across implicates. The final standard error is pooled across implicates using Reiter's pooling rules for partially synthetic data (Reiter, 2003). The estimate and standard error of each implicate is calculated by \code{\link[survey]{svyby}} using the provided observation weights and sample design.
}
\examples{
# Build a fusion model using RECS microdata
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fit <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Generate 5 implicates of synthetic 'fusion.vars',
#  using original RECS data as the recipient
recipient <- recs[predictor.vars]
sim <- fuseM(data = recipient, train.object = fit, M = 5)

# Specify the survey design of the recipient microdata
# Settings are survey-specific; see page 9 here:
# https://www.eia.gov/consumption/residential/data/2015/pdf/microdata_v3.pdf
recs.design <- survey::svrepdesign(weights = recs$weight,
                                   type = "Fay",
                                   rho = 0.5,
                                   repweights = select(recs, starts_with("rep_")),
                                   mse = TRUE)

# Analyze electricity consumption, by climate zone
result <- analyze(formula = electricity ~ 1,
                  by = "climate",
                  synthetic = sim,
                  static = recipient,
                  design = recs.design,
                  donor.N = nrow(recs))
head(result)
}
\references{
Reiter, J.P. (2003). Inference for Partially Synthetic,
Public Use Microdata Sets. \emph{Survey Methodology}, \bold{29}, 181-189.
}
