% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse}
\alias{fuse}
\title{Fuse variables to a recipient dataset}
\usage{
fuse(data, train.object)
}
\arguments{
\item{data}{Data frame. Recipient dataset. All categorical variables should be factors and ordered whenever possible. Data types and levels are strictly validated against predictor variables defined in \code{train.object}.}

\item{train.object}{Output from successfull call to \link{train}.}
}
\value{
A data frame with same number of rows as \code{data} and one column for each synthetic fusion variable defined in \code{train.object}. The order of the columns reflects the order in which they where fused.
}
\description{
Fuse variables to a recipient dataset.
}
\examples{
donor <- recs
recipient <- subset(recs, select = c(division, urban_rural, climate, income, age, race))
fusion.vars <- setdiff(names(donor), names(recipient))
fit <- train(data = donor, y = fusion.vars)
sim <- fuse(data = recipient, train.object = fit)
}
