% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse}
\alias{fuse}
\alias{fuseM}
\title{Fuse variables to a recipient dataset}
\usage{
fuse(
  data,
  file,
  k = 5,
  max_dist = 0,
  idw = FALSE,
  cores = 1,
  ignore_nearest = FALSE,
  seed = NULL
)

fuseM(..., M, fun = fusionModel::fuse, seeds = NULL)
}
\arguments{
\item{data}{Data frame. Recipient dataset. All categorical variables should be factors and ordered whenever possible. Data types and levels are strictly validated against predictor variables defined in \code{model}.}

\item{file}{Character. Path to fusion model file (.fsn) generated by successful call to \link{train}.}

\item{k}{Integer. Number of nearest neighbors to identify among the donor observations.}

\item{max_dist}{Numeric. Controls the maximum allowable distance when identifying up to \code{k} nearest neighbors. \code{max_dist = 0} (default) means no distance restriction is applied.}

\item{idw}{Logical. Should inverse distance weighting be used when randomly selecting a donor observation from the \code{k} nearest neighbors?}

\item{cores}{Integer. Number of cores used. Only applicable on Unix systems.}

\item{ignore_nearest}{Logical. If \code{TRUE}, nearest neighbor in KNN search is dropped; useful for validation exercises.}

\item{seed}{Set random seed if needed.}

\item{...}{Arguments passed to \code{fuse()}.}

\item{M}{Integer. Number of implicates to simulate.}
}
\value{
For \code{fuse()}, a data frame with same number of rows as \code{data} and one column for each synthetic fusion variable. The order of the columns reflects the order in which they where fused.

For \code{fuseM()}, a data frame with number of rows equal to \code{M * nrow(data)}. Integer column ".M" indicates implicate assignment of each observation. Note that the ordering of recipient observations is consistent within implicates, so do not change the row order if using with \code{analyze()}.
}
\description{
Fuse variables to a recipient dataset using a .fsn model produced by \link{train} using conditional distribution matching. \code{fuseM()} provides a convenience wrapper for generating multiple implicates. Output of \code{fuseM()} can be passed to \code{analyze()}.
}
\details{
For each record in \code{data}, the predicted conditional distribution values are used to identify the \code{k} most-similar observations in the original donor data along the same dimensions. One of the \code{k} nearest neighbors is randomly selected to donate its observed/"real" response value for the fusion variable(s) in question. The random selection uses inverse distance weighting if \code{idw = TRUE}.

The \code{max_dist} value is a relative scaling factor. The search radius passed to \code{\link[RANN]{nn2}} is \code{max_dist * medDist}, where \code{medDist} is the median pairwise distance calculated among all of the donor observations. This approach allows \code{max_dist} to adjust the search radius in a consistent way across all fusion variables/blocks.
}
\examples{
# Build a fusion model using RECS microdata
# Note that "test_model.fsn" will be written to working directory
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
train(data = recs, y = fusion.vars, x = predictor.vars, file = "test_model.fsn")

# Generate single implicate of synthetic 'fusion.vars',
#  using original RECS data as the recipient
recipient <- recs[predictor.vars]
sim <- fuse(data = recipient, file = "test_model.fsn")
head(sim)

# Calling fuse() again produces different results
sim <- fuse(data = recipient, file = "test_model.fsn")
head(sim)

# Generate multiple implicates via fuseM()
# This can be passed to 'implicates' argument in ?analyze()
sim <- fuseM(data = recipient, file = "test_model.fsn", M = 5)
head(sim)
table(sim$M)
}
