% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionModel-package.R
\docType{package}
\name{fusionModel-package}
\alias{fusionModel}
\alias{fusionModel-package}
\title{fusionModel: Fuse Variables from One Dataset to Another}
\description{
fusionModel enables variables unique to a "donor" dataset to be statistically simulated for (i.e. fused to) a "recipient" dataset.
}
\author{
\strong{Maintainer}: Kevin Ummel \email{ummel@berkeley.edu}

}
\keyword{internal}
