% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Extract predictor variable importance from a fusion model}
\usage{
importance(fsn)
}
\arguments{
\item{fsn}{Character. Path to fusion model file (.fsn) generated by \code{\link{train}}.}
}
\value{
A named list containing \code{detailed} and \code{summary} importance results. The \code{summary} results are most useful, as they return the average importance of each predictor across potentially multiple underlying LightGBM models; i.e. zero ("z"), mean ("m"), or quantile ("q") models. See Examples for suggested plotting of results.
}
\description{
Returns predictor variable (feature) importance of underlying LightGBM models stored in a fusion model file (.fsn) on disk.
}
\details{
Importance metrics are computed via \code{\link[lightgbm]{lgb.importance}}. Three types of measures are returned; "gain" is typically the preferred measure.
}
\examples{
# Build a fusion model using RECS microdata
# Note that "fusion_model.fsn" will be written to working directory
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fsn.path <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Extract predictor variable importance
ximp <- importance(fsn.path)

# Plot summary results
library(ggplot2)
ggplot(ximp$summary, aes(x = x, y = gain)) +
  geom_bar(stat = "identity") +
  facet_grid(~ y) +
  coord_flip()

# View detailed results
View(ximp$detailed)
}
