% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonic.R
\name{monotonic}
\alias{monotonic}
\title{Create a monotonic relationship between two variables}
\usage{
monotonic(
  x,
  y,
  w = NULL,
  preserve = TRUE,
  expend = TRUE,
  fast = TRUE,
  nmax = 5000,
  plot = FALSE
)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}

\item{w}{Numeric. Optional observation weights.}

\item{preserve}{Logical. Preserve the original mean of the \code{y} values in the returned values?}

\item{expend}{Logical. Assume \code{y} is an expenditure variable? If \code{TRUE}, a safety check is implemented to ensure \code{y > 0} when \code{x = 0}.}

\item{fast}{Logical. If \code{TRUE}, only \code{\link[scam]{supsmu}} is used with coercion of result to monotone.}

\item{nmax}{Integer. Maximum number of observations to use for smoothing. Set lower for faster computation. \code{nmax = Inf} eliminates sampling.}

\item{plot}{Logical. Plot the (sampled) data points and derived monotonic relationship?}
}
\value{
A numeric vector of modified \code{y} values. Optionally, a plot showing the returned monotonic relationship.
}
\description{
\code{monotonic()} returns modified values of input vector \code{y} that are smoothed, monotonic, and consistent across all values of input \code{x}. It was designed to be used post-fusion when one wants to ensure a plausible relationship between consumption (\code{x}) and expenditure (\code{y}), under the assumption that all consumers face an identical, monotonic pricing structure. By default, the mean of the returned values is forced to equal the original mean of \code{y} (\code{preserve = TRUE}). The direction of monotonicity (increasing or decreasing) is detected automatically, so use cases are not limited to consumption and expenditure variables.
}
\details{
The initial smoothing is accomplished via \code{\link[scam]{supsmu}} with the result coerced to monotone. If \code{fast = FALSE} and the coercion step modifies the values too much, a second smooth is attempted via a \code{\link[scam]{scam}} model with either a monotone increasing or decreasing constraint. If the SCAM fails to fit, the function falls back to \code{\link[stats]{lm}} with simple linear predictions. If \code{y = 0} when \code{x = 0} (as typical for consumption-expenditure variables), then that outcome is enforced in the result. The input data are randomly sampled to no more than \code{nmax} observations, if necessary, for speed.
}
\examples{
y <- monotonic(x = recs$propane_btu, y = recs$propane_expend, plot = TRUE)
mean(recs$propane_expend)
mean(y)
}
