% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonic.R
\name{monotonic}
\alias{monotonic}
\title{Ensure a monotonic relationship between two variables}
\usage{
monotonic(
  x,
  y,
  w = NULL,
  preserve_mean = TRUE,
  preserve_type = TRUE,
  plot = FALSE
)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}

\item{w}{Numeric. Optional observation weights.}

\item{preserve_mean}{Logical. Preserve the original mean of the \code{y} values in the returned values?}

\item{preserve_type}{Logical. Preserve the original data type of the \code{y} values in the returned values?}

\item{plot}{Logical. Plot the (sampled) data points and derived monotonic relationship?}
}
\value{
A numeric vector of modified \code{y} values. Optionally, a plot showing the returned monotonic relationship.
}
\description{
\code{monotonic()} returns modified values of input vector \code{y} that are smoothed, monotonic, and consistent across all values of input \code{x}. It was designed to be used post-fusion when one wants to ensure a plausible relationship between consumption (\code{x}) and expenditure (\code{y}), under the assumption that all consumers face an identical, monotonic pricing structure. By default, the mean of the returned values is forced to equal the original mean of \code{y} (\code{preserve_mean = TRUE}). The direction of monotonicity (increasing or decreasing) is detected automatically, so use cases are not limited to consumption and expenditure variables.
}
\details{
The initial smoothing is accomplished via \code{\link[scam]{supsmu}} with the result coerced to monotone. If the coercion step modifies the values too much, a second smooth is attempted via a \code{\link[scam]{scam}} model with either a monotone increasing or decreasing constraint. If the SCAM fails to fit, the function falls back to \code{\link[stats]{lm}} with simple linear predictions. If \code{y = 0} when \code{x = 0} (as typical for consumption-expenditure variables), then that outcome is enforced in the result. The input data are randomly sampled to no more than 10,000 observations, if necessary, for speed.
}
\examples{
y <- monotonic(x = recs$propane_btu, y = recs$propane_expend, plot = TRUE)
mean(recs$propane_expend)
mean(y)
}
