% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fsd.R
\name{read_fsd}
\alias{read_fsd}
\title{Read fusion output from disk}
\usage{
read_fsd(fsd, cores = data.table::getDTthreads())
}
\arguments{
\item{fsd}{Character. File path ending in \code{.fsd} as produced by call to \code{\link{fuse}}.}

\item{cores}{Integer. Number of cores used by \code{\link[data.table]{fread}}.}
}
\value{
A \code{\link[data.table]{data.table}} with integer column "M" indicating the implicate assignment of each observation. Note that the ordering of recipient observations is consistent within implicates, so do not change the row order if using with \code{\link{analyze}} or \code{\link{validate}}.
}
\description{
Read fusion output that was written directly to disk via \code{\link{fuse}}.
}
\details{
The file written by \code{\link{fuse}} includes specially formatted metadata in the column names that \code{read_fsd} uses to construct the appropriate output values and classes. Reading \code{fsd} with a different file reader function will not give correct results.
}
\examples{
# Build a fusion model using RECS microdata
# Note that "fusion_model.fsn" will be written to working directory
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fsn.path <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Write fusion output directly to disk
# Note that "results.fsd" will be written to working directory
recipient <- recs[predictor.vars]
sim <- fuse(data = recipient, fsn = fsn.path, M = 5, csv = "results.fsd")

# Read the fusion output saved to disk
sim <- read_fsd(sim)
head(sim)

}
