% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a fusion model on donor data}
\usage{
train(
  data,
  y,
  x = NULL,
  ignore = NULL,
  weight = NULL,
  mc = FALSE,
  lasso = NULL,
  maxcats = 10,
  ...
)
}
\arguments{
\item{data}{Data frame. Donor dataset. All categorical variables should be factors and ordered whenever possible.}

\item{y}{Character vector. The variables to fused to recipient dataset. Can include \link[base:regex]{regular expressions}.}

\item{x}{Character vector. Predictor variables common to donor and recipient. Can include regular expressions. If NULL, all variables other than those in \code{y} and \code{weight} are used. Only one of \code{x} and \code{ignore} can be non-NULL.}

\item{ignore}{Character vector. Alternative way to specify predictor variables. Can include regular expressions. If non-NULL, all variables other than those in \code{y}, \code{weight}, and \code{ignore} are used. Only one of \code{x} and \code{ignore} can be non-NULL.}

\item{weight}{Character vector. Name of the observation weights column. If NULL (default), uniform weights are assumed.}

\item{mc}{Logical. Should multicore processing be used? Implementation uses \code{\link[parallel]{mclapply}}, so \code{mc = TRUE} will fail on Windows.}

\item{lasso}{Numeric (0-1) or NULL. Controls extent of (fast) predictor variable pre-screening via LASSO regression. If NULL (default), no pre-screening is performed. See Details.}

\item{maxcats}{Positive integer. Maximum number of levels allowed in an unordered factor predictor variable when the response (fusion) variable is also an unordered factor. Prevents excessive \code{\link[rpart]{rpart}} computation time. A K-means clustering strategy is used to cluster the predictor to no more than \code{maxcats} levels.}

\item{...}{Optional arguments passed to \code{\link[rpart]{rpart}} to control tree-building. By default \code{cp = 0}, \code{xval = 0}, \code{minbucket = 50} (\code{minbucket = 10} for discrete models), and all other arguments are left at default values.}
}
\value{
A list containing trained model information to be passed to \link{fuse}.
}
\description{
Train a fusion model on donor data.
}
\details{
When \code{lasso} is non-NULL, predictor variables are "pre-screened" using LASSO regression via \code{\link[rpart]{glmnet}} prior to fitting a \code{\link[rpart]{rpart}} model. Predictors with a LASSO coefficient of zero are excluded from consideration. This can speed up tree-fitting considerably when \code{data} is large. Lower values of \code{lasso} are more aggressive at excluding predictors; the LASSO \emph{lambda} is chosen such that the deviance explained is at least \code{lasso}-\% of the maximum. To ensure the LASSO step is fast, pre-screening is only used for numeric and ordered factor response variables. No predictor pre-screening occurs when the response is an unordered factor.
}
\examples{
donor <- recs
recipient <- subset(recs, select = c(division, urban_rural, climate, income, age, race))
fusion.vars <- setdiff(names(donor), names(recipient))
fit <- train(data = donor, y = fusion.vars)
}
